Fix critical bugs and improve security in BAP firmware
- Fix HTTP chunked encoding parser for proper API response handling
- Fix NTP time sync to detect actual time changes (was using stale 2024 time)
- Add WiFi connection retry logic with 5 attempts and 30s timeout
- Migrate secrets (WiFi password, API key) from SPIFFS to NVS for security
- Replace weak XOR checksum with Fletcher-16 in mesh protocol
- Switch API calls from HTTP to HTTPS for encrypted transit
- Fix timezone handling with automatic DST via configTzTime()
- Extract parseVisitArray() helper to reduce code duplication
- Add input validation for stop_id (4-7 digits)
- Remove unused BAPConfigExt struct

+++ examples/bap_firmware/MyMesh.h
@@ -67,8 +67,8 @@ struct BAPArrivalPacket {
   uint16_t sequence;        // 2 bytes - Sequence number for dedup
   uint32_t generated_at;    // 4 bytes - When gateway fetched this data
   BusArrival arrivals[5];   // 155 bytes - Up to 5 arrivals
-  uint8_t checksum;         // 1 byte - Simple checksum
-};                          // Total: 164 bytes
+  uint16_t checksum;        // 2 bytes - Fletcher-16 checksum
+};                          // Total: 165 bytes
 
 // BAP Configuration stored in SPIFFS
 struct BAPConfig {
@@ -138,5 +138,5 @@ protected:
   bool allowPacketForward(const mesh::Packet* packet) override;
 
 private:
-  uint8_t calcChecksum(const uint8_t* data, size_t len);
+  uint16_t calcChecksum(const uint8_t* data, size_t len);
 };
